add:添加权限表,详见sql-readme.md 登录功能测试通过

+++ common/utils/src/main/java/com.nxftl.doc.common.util/util/Token.java
@@ -53,8 +53,8 @@ public class Token {
      * @param decodedJWT
      * @return
      */
-    public static Integer getUserId(DecodedJWT decodedJWT){
-        return decodedJWT.getClaim(Config.USER_ID).asInt();
+    public static Long getUserId(DecodedJWT decodedJWT){
+        return decodedJWT.getClaim(Config.USER_ID).asLong();
     }
